SDK - Move account-manager.ts into the root directory
