+++ sdk/script/account-manager.ts
